Abstract Jobs and move to diffbot_lib
