+++ crates/diffbot_lib/Cargo.toml
@@ -5,7 +5,10 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
+serde_json = { version = "1.0" }
 octocrab = { git = "https://github.com/XAMPPRocky/octocrab" }
 anyhow = "1.0.57"
 derive_builder = "0.11.2"
 chrono = "0.4.19"
+flume = "0.10.12"
+tokio = "1.18.1"
