+++ crates/diffbot_lib/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod git;
 pub mod github;
+pub mod job;
